requirements.txt
boto3
pandas
pyarrow
yfinance
pyyaml
numpy
gymnasium
stable_baselines3